+++ lib/rails/generators/mongoid_generator.rb
@@ -10,15 +10,6 @@ module Mongoid
         @_mongoid_source_root ||=
           File.expand_path("../#{base_name}/#{generator_name}/templates", __FILE__)
       end
-
-      unless methods.include?(:module_namespacing)
-
-        # This is only defined on Rails edge at the moment, so include here now
-        # as per: https://github.com/mongoid/mongoid/issues/744
-        def module_namespacing(&block)
-          yield if block
-        end
-      end
     end
 
     class ActiveModel < ::Rails::Generators::ActiveModel